Fix types

+++ src/projects/gql/delete-project.gql.ts
@@ -1,7 +1,7 @@
 import { gql } from '@apollo/client'
 
 export const deleteProjectMutation = gql`
-  mutation DeleteProject($id: Float!) {
+  mutation DeleteProject($id: Int!) {
     deleteProject(id: $id) {
       affected
     }
